+++ tool_graph/graph_model.py
@@ -71,24 +71,6 @@ class LocationTimeseriesModel(BaseModel):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.dataChanged.connect(self.update_labels)
-
-    def update_labels(self, index):
-        if self.columns[index.column()].name == "object_label":
-            # Update user-editable label of other plots for same feature
-            updated_item = self.rows[index.row()]
-            updated_feature_id = updated_item.object_id.value
-            updated_grid = updated_item.result.value.parent()
-            updated_object_type = updated_item.object_type.value
-            assert updated_grid
-            new_label = updated_item.object_label.value
-            for item in self.rows:
-                if (item.object_id.value == updated_feature_id and
-                        item.result.value.parent() is updated_grid and
-                        item.object_type.value == updated_object_type):  # e.g. flowline
-                    self.blockSignals(True)
-                    item.object_label.value = new_label
-                    self.blockSignals(False)
 
     def get_color(self, idx: int) -> QColor:
         if not self.feature_color_map: